role permission some done

+++ routes/web.php
@@ -30,13 +30,13 @@ Route::prefix('dashboard')->name('dashboard.')->group(function(){
     //Dashboard index route:
     Route::get('/', [BackendController::class, 'index'])->name('dashboard');
 
-    Route::controller(RolePermissionController::class)->middleware(['role:super-admin'])->prefix('role')->name('role.')->group(function(){
-        Route::get('/','index')->name('index');
-        Route::get('create','create')->name('create');
-        Route::post('store','store')->name('store');
-        Route::get('edit/{id}','edit')->name('edit');
-        Route::put('update/{id}','update')->name('update');
-        Route::delete('delete/{id}','destroy')->name('delete');
+    Route::controller(RolePermissionController::class)->prefix('role')->name('role.')->group(function(){
+        Route::get('/','index')->middleware('role_or_permission:super-admin|show')->name('index');
+        Route::get('create','create')->middleware('role_or_permission:super-admin')->name('create');
+        Route::post('store','store')->middleware('role_or_permission:super-admin|admin')->name('store');
+        Route::get('edit/{id}','edit')->middleware('role_or_permission:super-admin|admin')->name('edit');
+        Route::put('update/{id}','update')->middleware('role_or_permission:super-admin|admin')->name('update');
+        Route::delete('delete/{id}','destroy')->middleware('role_or_permission:super-admin')->name('delete');
     });
 
      Route::controller(PermissionController::class)->middleware(['role:super-admin'])->prefix('permission')->name('permission.')->group(function(){
